Use package constants in key expansion module.
Use package constants instead of manually typed values throughout the
key expansion module design.

Initialize i and j to 0. Testbench was failing because i and j were left
uninitialized and the simulator used random values which were outside of
the range of the arrays L and S. On the first clock cycle, i and j are
reset to 0 (state "idle") so initializing them doesn't compromise the
design.